Restructure and generalize the Puppet manifests.
This isn't quite working yet, but I did a lot of work to
generalize the way Puppet was set up and make it work with
broker and stunnel. I also externalized many of the options
so that users can customize some of the behavior without
having to modify the base manifest.